requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.6
django-annoying==0.10.6
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-celery-results==2.4.0
django-codemirror2==0.2
django-cte==1.1.4
django-dramatiq==0.11.2
django-environ==0.10.0
django-filter==22.1
django-js-asset==2.0.0
django-jsonstore==0.5.1
django-mptt==0.14.0
django-polymorphic==3.1.0
django-render-block==0.9.2
django-reversion==5.0.4
django-tables2==2.5.1
django-templated-email==3.0.1
django-tree-queries==0.14.0
django-viewflow==1.11.0
dramatiq==1.14.1
gevent==22.10.2
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
joeflow==1.2.2
kombu==5.2.4
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycparser==2.21
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==5.5.2
